+++ src/components/editor/RowView.svelte
@@ -4,7 +4,6 @@
     import type Row from '@nodes/Row';
     import NodeSequenceView from './NodeSequenceView.svelte';
     import NodeView from './NodeView.svelte';
-
     export let node: Row;
 </script>