+++ src/nodes/Row.ts
@@ -1,6 +1,5 @@
 import type { Grammar, Replacement } from './Node';
 import Token from './Token';
-import Bind from './Bind';
 import Expression from './Expression';
 import Glyphs from '../lore/Glyphs';
 import Purpose from '../concepts/Purpose';
@@ -15,13 +14,18 @@ import ValueException from '../values/ValueException';
 import StructureValue from '../values/StructureValue';
 import { TABLE_CLOSE_SYMBOL, TABLE_OPEN_SYMBOL } from '../parser/Symbols';
 import type Locales from '../locale/Locales';
+import Input from './Input';
 
 export default class Row extends Node {
     readonly open: Token;
-    readonly cells: Expression[];
+    readonly cells: (Input | Expression)[];
     readonly close: Token | undefined;
 
-    constructor(open: Token, cells: Expression[], close: Token | undefined) {
+    constructor(
+        open: Token,
+        cells: (Input | Expression)[],
+        close: Token | undefined,
+    ) {
         super();
 
         this.open = open;
@@ -35,7 +39,7 @@ export default class Row extends Node {
         return new Row(
             new Token(TABLE_OPEN_SYMBOL, Sym.TableOpen),
             cells,
-            new Token(TABLE_CLOSE_SYMBOL, Sym.TableClose)
+            new Token(TABLE_CLOSE_SYMBOL, Sym.TableClose),
         );
     }
 
@@ -52,7 +56,7 @@ export default class Row extends Node {
                     node(Sym.Select),
                     node(Sym.Insert),
                     node(Sym.Delete),
-                    node(Sym.Update)
+                    node(Sym.Update),
                 ),
             },
             { name: 'cells', kind: list(true, node(Expression)), space: true },
@@ -64,7 +68,7 @@ export default class Row extends Node {
         return new Row(
             this.replaceChild('open', this.open, replace),
             this.replaceChild('cells', this.cells, replace),
-            this.replaceChild('close', this.close, replace)
+            this.replaceChild('close', this.close, replace),
         ) as this;
     }
 
@@ -72,12 +76,12 @@ export default class Row extends Node {
         return Purpose.Bind;
     }
 
-    allBinds() {
-        return this.cells.every((cell) => cell instanceof Bind);
+    getDependencies() {
+        return [...this.cells];
     }
 
     allExpressions() {
-        return this.cells.every((cell) => !(cell instanceof Bind));
+        return this.cells.every((cell) => !(cell instanceof Input));
     }
 
     computeConflicts() {
@@ -98,13 +102,13 @@ export function getRowFromValues(
     evaluator: Evaluator,
     creator: EvaluationNode,
     table: TableType,
-    values: Value[]
+    values: Value[],
 ) {
     const evaluation = new Evaluation(
         evaluator,
         creator,
         table.definition,
-        evaluator.getCurrentEvaluation()
+        evaluator.getCurrentEvaluation(),
     );
 
     for (let c = 0; c < table.columns.length; c++) {